requirements.txt
speechrecognition
pygame
pyaudio
gtts
requests
openai
python-dotenv
AppOpener
ttkbootstrap